fix(ci): split oversized diplomacy and scoring tests for repo lint (#2509)
Unblocks PR #2552 Quality gate: logic test file >400 lines and AI test
file >1000 non-comment lines.
